arm64: add basic DTS for i.MX8MQ
This adds the basic DTS for the i.MX8MQ.
For now only the following peripherals are supported:
- IOMUXC (pin controller)
- CCM (clock controller)
- GPIO
- UART
- uSDHC (SD/eMMC controller)
- FEC (ethernet controller)
- i2c

This is enough to get a very basic board support up and running.

One known limitation is that the driver for the GPC interrupt
controller is still missing, rendering the CPU sleep states unusable
as there is nothing waking them up anymore. This will be fixed in
due course.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
---
v2:
 - move to freescale folder
 - change compatibles to "fsl" vendor prefix
 - remove unnecessary newlines
 - move status property to be the last in all node
 - remove imx21 compatibles from UART nodes
 - document compatible